Adjusted the defualt settings main tag
Had to adjust the main tag default settings due
to the header being sticky to the top of the web
page.

used the calc function to take 100vh and minus
that from the 60px heigh of the header which can
can be found in the header.css file.

Added the margin to also push it down by 60px